backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
email-validator
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==3.2.2